refactor: __init__.py God Object 분할 — pipeline_executor + report_renderer 추출
__init__.py에서 2개 모듈을 추출하여 SRP 적용:

- pipeline_executor.py (642줄): 14개 파이프라인 실행/렌더링 함수
- report_renderer.py (258줄): 4개 리포트 생성 함수 + 5개 상수

__init__.py: 2,554줄 → 1,768줄 (-786줄, -31%)
순환 import: deferred import로 해결
전체 테스트: 2972 passed
